完成支付扣减接口

+++ user-service/src/main/java/com/harbor/secondHand/user/domain/dto/UserMessageDTO.java
@@ -15,7 +15,7 @@ public class UserMessageDTO {
     private String avatar;
     private String phone;
     private String email;
-    private Integer balance;
+    private BigDecimal balance;
     private Integer creditScore;
     private LocalDateTime updateTime;
     private String operationType;
